+++ CHANGES.md
@@ -1,6 +1,7 @@
 # Release notes
 All notable changes to this project will be documented in this file.
 
+## 0.5.0
 - Added `Technical reports` molecule
 - Created `menu` molecule